requirements.txt
python-telegram-bot==20.7
asyncio==3.4.3
postgrest-py==0.10.6
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiohttp==3.8.6 